requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==7.0.0
httpx==0.25.2
python-dotenv==1.0.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest-playwright==0.4.3

